+++ library/src/compute/mod.rs
@@ -4,7 +4,6 @@ pub mod calc;
 mod construct;
 mod data;
 mod foundations;
-mod to_base;
 
 pub use self::construct::*;
 pub use self::data::*;
